+++ tests/test_refdata.py
@@ -123,7 +123,7 @@ class TestRefdata(unittest.TestCase):
                                               host=biosqlhost,
                                               db=biosqldb,
                                               port=biosqlport)
-        refdata = ReferenceData(server=server, verbose=True)
+        refdata = ReferenceData(server=server)
         self.assertIsInstance(refdata, ReferenceData)
         self.assertTrue(refdata.server_avail)
         self.assertFalse(refdata.imgtdat)
@@ -139,7 +139,7 @@ class TestRefdata(unittest.TestCase):
                                               db=biosqldb,
                                               port=biosqlport)
         for db in self.dblist:
-            refdata = ReferenceData(server=server, dbversion=db, verbose=True)
+            refdata = ReferenceData(server=server, dbversion=db)
             self.assertEqual(refdata.dbversion, db)
             self.assertTrue(refdata.server_avail)
             self.assertFalse(refdata.imgtdat)
@@ -154,7 +154,7 @@ class TestRefdata(unittest.TestCase):
                                               host=biosqlhost,
                                               db=biosqldb,
                                               port=biosqlport)
-        refdata = ReferenceData(server=server, verbose=True)
+        refdata = ReferenceData(server=server)
         input_seq = self.data_dir + '/exact_seqs.fasta'
         self.assertFalse(refdata.imgtdat)
 
@@ -191,7 +191,7 @@ class TestRefdata(unittest.TestCase):
                                               host=biosqlhost,
                                               db=biosqldb,
                                               port=biosqlport)
-        refdata1 = ReferenceData(server=server, verbose=True)
+        refdata1 = ReferenceData(server=server)
         refdata2 = ReferenceData()
 
         datseqs = [a for a in refdata2.imgtdat
@@ -221,7 +221,7 @@ class TestRefdata(unittest.TestCase):
                                               host=biosqlhost,
                                               db=biosqldb,
                                               port=biosqlport)
-        refdata = ReferenceData(server=server, alignments=True, verbose=True)
+        refdata = ReferenceData(server=server, alignments=True)
         for ex in self.expected['align']:
             locus = ex['locus']
             allele = ex['name'].split("_")[0]